third_party/blink/web_tests/accessibility/color-well-expected.txt

CONSOLE WARNING: The specified value "notacolor" does not conform to the required format.  The value must be a valid CSS color.
This test ensures that the correct accessibility properties can be retrieved from input elements of type color.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS axColorwell.role is "AXRole: AXColorWell"
PASS axColorwell.colorValue is 4278190080
PASS axColorwell.stringValue is "AXValue: rgba(0, 0, 0, 1)"
PASS axColorwell.role is "AXRole: AXColorWell"
PASS axColorwell.colorValue is 4294901760
PASS axColorwell.stringValue is "AXValue: rgba(255, 0, 0, 1)"
PASS axColorwell.role is "AXRole: AXColorWell"
PASS axColorwell.colorValue is 4278190080
PASS axColorwell.stringValue is "AXValue: rgba(0, 0, 0, 1)"
PASS successfullyParsed is true

TEST COMPLETE

  
